Refactor Main component; import in index.jsx
