+++ webpack.config.js
@@ -2,7 +2,7 @@ const webpack = require('webpack');
 const path = require('path');
 
 const config = {
-  entry: path.resolve(__dirname, 'static/js/components/Main.jsx'),
+  entry: path.resolve(__dirname, 'static/js/index.jsx'),
   output: {
     path: path.resolve(__dirname, 'static/build'),
     filename: 'bundle.js'
